[DATAVIC-946] refactor xloader integration by creating a new IAR xloader plugin
- Updated setup.py to reference the new IAR xloader plugin.
- Introduced a new plugin (iar_xloader_plugin.py) that extends xloader functionality to handle new-resource-only submissions during dataset updates.
- Removed the previous xloader implementation from plugins.py to streamline the codebase.
