+++ setup.py
@@ -28,7 +28,7 @@ setup(
         [ckan.plugins]
         datavicmain_dataset = ckanext.datavicmain.plugins:DatasetForm
         datavicmain_home = ckanext.datavicmain_home.plugins:DatavicHomePlugin
-        datavic_iar_xloader = ckanext.datavicmain.plugins:DatavicIARXLoaderPlugin
+        datavic_iar_xloader = ckanext.datavicmain.iar_xloader_plugin:DatavicIARXLoaderPlugin
 
 
         [fanstatic.libraries]
